% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunTEST.R
\name{dunnTEST}
\alias{dunnTEST}
\title{A function for non-parametric multi-group test using dunn.test package}
\usage{
dunnTEST(data, method = "bh")
}
\arguments{
\item{data}{a data frame with grouping information}

\item{method}{adjusts the p-value for multiple comparisons}
}
\value{
a data frame
}
\description{
A function for non-parametric multi-group test using dunn.test package
}
\author{
Zhonghui Gai
}
