% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggalluvial.abun.R
\name{ggalluvial.abun}
\alias{ggalluvial.abun}
\title{Draw a alluvim plot using ggplot2 and ggalluvial package}
\usage{
ggalluvial.abun(data)
}
\arguments{
\item{data}{a data.frame abtained from summaryAll function}
}
\value{

}
\description{
Draw a alluvim plot using ggplot2 and ggalluvial package
}
\examples{
phylum <- read.csv("phylum.csv")
data <- phylum[, 1:7]
results <- data.frame(group = data$group, data[, rev(colnames(data)[-1])])
results <- summaryAll(results)
results$mean <- 100*results$mean
ggalluvial.abun(results) + theme(legend.position = "top") +
ggsci::scale_fill_aaas()
}
\author{
Zhonghui Gai
}
