% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggboxplot.Let.Sig.dunnTEST.R
\name{ggboxplot.Let.Sig.dunTEST}
\alias{ggboxplot.Let.Sig.dunTEST}
\title{A function for multiple group ggplot2 based boxplot using dunn.test}
\usage{
ggboxplot.Let.Sig.dunTEST(
  data,
  adjust = FALSE,
  set.levels = TRUE,
  letterMark = TRUE,
  unites = c("IS\\n(g/L)", "PCS\\n(g/L)", "TMAO\\n(g/L)", "3AA\\n(g/L)")
)
}
\arguments{
\item{data}{the data frame containing the grouping information}

\item{adjust}{adjust the p.value}

\item{set.levels}{boolean value for facet title}

\item{letterMark}{a boolean value to show LETTERS or not}

\item{unites}{the new levels for facet title text, usually containing two lines}
}
\value{
a ggplot boxplot
}
\description{
A function for multiple group ggplot2 based boxplot using dunn.test
}
\examples{
data <- iris
colnames(data)[5] <- "group"
ggboxplot.Let.Sig.dunTEST(data = data, adjust = FALSE,
set.levels = TRUE, letterMark = TRUE,
unites = c("Sepal Length\n(g/L)", "Sepal Width\n(g/L)",
           "Petal Length\n(g/L)", "Petal Width\n(g/L)"))
}
\author{
Zhonghui Gai
}
