% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggboxplot.Let.Sig.R
\name{ggboxplot.LetSig}
\alias{ggboxplot.LetSig}
\title{Draw boxplot of multiple variable using grouping info. The signicant info is
shown by Letters. Function3}
\usage{
ggboxplot.LetSig(data)
}
\arguments{
\item{data}{A dataframe containing grouping info and multiple numeric variables.}
}
\value{
a boxplot faceted by different variables.
}
\description{
Draw boxplot of multiple variable using grouping info. The signicant info is
shown by Letters. Function3
}
\examples{
data <- iris
colnames(data)[5] <- "group"
ggboxplot.LetSig(data = data) + theme(legend.position = 0)
}
\author{
Zhonghui Gai
}
