% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggboxplot.LetSig1.R
\name{ggboxplot.LetSig1}
\alias{ggboxplot.LetSig1}
\title{Similar function of ggboxplot.LetSig, only for one variable}
\usage{
ggboxplot.LetSig1(data, facet = FALSE)
}
\arguments{
\item{data}{A dataframe containing grouping info and only one numeric variables.}
}
\value{

}
\description{
Similar function of ggboxplot.LetSig, only for one variable
}
\examples{
data <- iris
colnames(data)[5] <- "group"
ggboxplot.LetSig1(data = data[, 4:5]) + theme(legend.position = 0)
}
\author{
Zhonghui Gai
}
