% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggboxplot_2grp.R
\name{ggboxplot_2grp}
\alias{ggboxplot_2grp}
\title{Boxplot based on ggplot2 of two groups and multiple numeric variables}
\usage{
ggboxplot_2grp(data, violin = TRUE, test = "wilcox.test", nrow = 1)
}
\arguments{
\item{data}{a data frame containing grouping information which only have two groups}

\item{violin}{boolean value to indicate whether to show violin plot}

\item{test}{the method used to compare, one of wilcox.test or t.test}

\item{nrow}{the number of rows}
}
\value{
a ggplot2 plot
}
\description{
Boxplot based on ggplot2 of two groups and multiple numeric variables
}
\examples{
data <- iris
colnames(data)[5] <- "group"
data <- data[1:100, ]  |> droplevels()
ggboxplot_2grp(data)
}
\author{
ZhonghuiGai
}
