% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpointplot.R
\name{ggpointplot}
\alias{ggpointplot}
\title{Draw a point&line plot using multiple variable according the grouping info.
Function 2.}
\usage{
ggpointplot(data, type = "simple", fun.data = mean_sdl, size = 4)
}
\arguments{
\item{data}{A dataframe containing the grouping info and more than one numeric
variables. The structure of data likes iris.}

\item{type}{The type of plot, one of "simple" and "facet", the default type
is "simple".}

\item{fun.data}{The function used to draw the errorbar, one of "mean_sdl" and
"mean_se"}

\item{size}{the size of point}
}
\value{
a point plot
}
\description{
Draw a point&line plot using multiple variable according the grouping info.
Function 2.
}
\examples{
library(ggplot2)
data <- iris
colnames(data)[5] <- "group"
ggpointplot(data = data, type = "facet", fun.data = mean_sdl) +
ggsci::scale_color_aaas() +
theme(legend.position = c(0.5, 0.1),
      legend.direction = "horizontal")
}
\author{
Zhonghui Gai
}
