% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryAll.R
\name{summaryAll}
\alias{summaryAll}
\title{Summarizes data}
\usage{
summaryAll(data)
}
\arguments{
\item{data}{A data frame containing the group collum for grouping info}
}
\value{
a data frame containing count, max, mean, standard deviation,
standard error of mean, 95% confidence interval
}
\description{
Summarizes data
}
\examples{
data <- iris
colnames(data)[5] <- "group"
summaryAll(data)
}
